start to split code in functions separate cli tool and node function, keep it modular
