+++ .eslintrc.js
@@ -5,6 +5,7 @@ module.exports = {
   },
   "extends": "plugin:prettier/recommended",
   "parserOptions": {
-    "ecmaVersion": 2017
+    "ecmaVersion": 2017,
+    "sourceType": "module"
   }
 };
