+++ examples/nicolas-hoizey.com/run.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 path=`pwd`
-node ../../index.js --contextsfile ./contexts.csv --url "file://${path}/page.html" --selector '.main img[srcset]:first-of-type' --maxviewport 2560 --delay 100 --verbose
+node ../../src/cli.js --contexts-file ./contexts.csv --url "file://${path}/page.html" --selector '.main img[srcset]:first-of-type' --max-viewport 2560 --delay 100 --verbose
